[Feat] 날씨 정보 구현-미완성/ [Modify] FileProcessor -> FileUtil 로 이름변경

+++ springboot/src/main/java/com/tave_app_1/senapool/entity/User.java
@@ -18,7 +18,7 @@ public class User {
 
     @Id @Column(name = "user_pk")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer userPK;
+    private Long userPK;
 
     @Column(name = "user_id",nullable = false)
     private String userId;
